Restructure tests to use new filenames and imports
