+++ tests/integration/translate-captions.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import 'dotenv/config';
-import { translateCaptions } from '../../src/translation';
+import { translateCaptions } from '../../src/functions';
 
 describe('Captions Translation Integration Tests', () => {
   const assetId = '88Lb01qNUqFJrOFMITk00Ck201F00Qmcbpc5qgopNV4fCOk';
